+++ test-resources/midje/semi_sweet_compile_out.clj
@@ -1,5 +1,5 @@
 
 ;; This file is used to check whether checks can be compiled out.
-(expect false => true)
+(expect (throw (Exception. "Should never be called"))  => true)
 (deftest tests-can-be-compiled-out
     (expect false => true))