+++ test-resources/midje/sweet_compile_out.clj
@@ -1,0 +1,5 @@
+
+;; This file is used to check whether facts can be compiled out.
+
+(fact
+  (throw (Exception. "This is supposed to compile out")))